requirements.txt
pytorch>=1.11.0+cpu
sklearn>=1.0.2
matplotlib>=3.5.1
